go/runtime/registry: Decouple host provisioner from runtime registry
